backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
httpx==0.28.1
asyncpg==0.31.0
python-dotenv==1.2.1
python-multipart==0.0.22
